+++ src/main/java/me/drex/instantfeedback/worldgen/FallenDarkOakTrunkPlacer.java
@@ -1,0 +1,4 @@
+package me.drex.instantfeedback.worldgen;
+
+public class FallenDarkOakTrunkPlacer {
+}